Attempt to fix more tests
